Move test_app and client from test_auth.py to separate fixture to re-use it later for extraction tests
